fixed product update img bug & added profile icon

+++ api/src/services/product.service.ts
@@ -102,8 +102,11 @@ const update = async (
 ): Promise<ProductDocument | null> => {
   let oldImage
   if (update.image) {
+    console.log('there is a new image')
+    console.log(update.image)
     const product = await Product.findOne({ slug: productName })
     oldImage = product && product.image
+    console.log(oldImage)
   }
 
   update.season = update.season.split(',') as Season[]
@@ -121,8 +124,8 @@ const update = async (
     { new: true }
   )
 
-  if (updatedProduct) {
-    if (oldImage) {
+  if (updatedProduct && oldImage) {
+    if (update.image !== oldImage) {
       removeFile(`images/products/${oldImage}`)
     }
   }
